feat: added form builder component
This commit adds a new component to the application that allows users to dynamically create and edit forms.

The form builder component is a drag-and-drop interface that allows users to add various form elements, such as text inputs, radio buttons, checkboxes, and textareas. Users can also set the label and value of each element, as well as add custom options for radio buttons and checkboxes.

Once a form is created, users can preview it in real-time to see how it will look and interact with users. The form builder also allows users to save the form as a JSON object, which can then be used to generate a fully functional HTML form.

This component is a valuable addition to the application as it provides a quick and easy way to create and edit forms without having to write any code. It is also a great tool for prototyping and testing new form designs.

- Added new  component to .
- Added new  page to .
- Updated  to include a link to the  page.
- Made various styling adjustments to the  component.
- Fixed a bug where the  component was not rendering correctly.