feat: add page to display list of in-charge authorities
Adds a new page to the dashboard that displays a list of in-charge authorities.

The page includes the following information for each authority:

- Name
- Designation
- Email
- Phone number
- Whether or not they have an appointment scheduled

Students can book an appointment with an authority if they have one scheduled.

This page allows students to easily contact and schedule appointments with the authorities they need to interact with.